Fix the sighandler in MainUI
MainUI patched to react to SIGTERM signal.

A sighandler to break out of the main loop was incorrectly implemented, set up to act on SIGKILL (9) which is not possible, as SIGKILL will always kill and leave the process no time to react.

The sighandler simply sets a value which is checked early in the main while(true) loop

This allows us to shut down MainUI gracefully by sending SIGTERM instead of SIGKILL which can cause all sorts of troubles, notably black screens.