Add wrapper for global GUI keyboard support.
Adds a keyboard wrapper library

We're now at the point where we need some level of graphical keyboard
within our own bins (for example, tweaks for tGeckos new menus for
about/wifi).

This PR will include the library for allow this, built from Aemii's C++
search filter source (Which means it's separate, PR for that repo to
follow).

The keyboard has to be a library as it's C++ but the wrapper works well,
it's fast and it returns the value cleanly, returning you back to the
tweaks menu stack once complete.

As part of this it also releases the A button in Tweaks event loop, if
you do not release this it stays pressed. When you return out of the
wrapper the button is still pressed, take this into account in your
event loops.

# How to use it in your feature (If it's in tweaks):

1. `#include "components/kbinput_wrapper.h"`
2. Call the wrapper function from `actions.h` that will return the
keyboards output, eg:

```c
const char* action_LaunchKeyboardWrapper(const char *initial_value, const char *title) {
    const char* result = launch_keyboard(initial_value, title);
    all_changed = true;
    return result;
}
```

3. On your menu item you'll need a void function so create a function to
call the wrapper:

```c
void getInputFromKeyboard() {
    const char *initial_value = "Enter here";  // The default entry in the textbox
    const char *title = "WiFi Key"; // The default title of the window
    const char* result = action_LaunchKeyboardWrapper(initial_value, title); // Call the keyboard
    printf("RESULT = %s\n", result); //Set result to your variable or whatever
    free((void*)result); // Free the result whenever (but don't forget)
}
```

4. Call it at your chosen location, for example on a tweaks menu stack
item:

```c
 list_addItem(&_menu_main,
                     (ListItem){
                         .label = "Keyboard",
                         .description = "noice",
                         .action = getInputFromKeyboard,
                         .icon_ptr = _get_menu_icon("tweaks_tools")
                     });
```

# How to use it in your feature (If it's not in tweaks):

1. Link the library in your makefile:

![image](https://github.com/OnionUI/Onion/assets/47260768/2763c72f-78f8-4da3-9cad-26daa5d146da)
2. Steps 1 -> 3 above but define your own functions. When calling the
keyboard, the function is:

![image](https://github.com/OnionUI/Onion/assets/47260768/6bfb4f01-be74-444c-ba0b-29bc129c458b)

# Example of keyboard called from menu item and output printed to
terminal:

![image](https://github.com/OnionUI/Onion/assets/47260768/2890ee6d-6e7d-4efb-a0d2-f2d57c389374)

![image](https://github.com/OnionUI/Onion/assets/47260768/bd39c7c1-4dff-446a-a5a5-2954cbf3e2e7)

https://github.com/OnionUI/Onion/assets/47260768/d6f44e12-0fe8-49c0-a9b1-dae19629d6b2

---------

Co-authored-by: XK9274 <XK9274@users.noreply.github.com>
Co-authored-by: Schmurtz <schmurtz@gmail.com>