Drastic integration
### Integration of Drastic:

_- The Game Switcher and Keymon now utilize Miyoo's recentlist.json
instead of Retroarch's content_history.lpl._
_- The system converts the old Game Switcher list to the new one._
_- Keymon has been modified to account for Drastic._
_- Scraper now supports NDS games._

**Game Switcher:**

- Retroarch games and those launched from the Emu folder (Type = 5) are
displayed.
- The old logic involved parsing the Retroarch game path + core and
reconstructing cmd_to_run.sh. The new logic calls the appropriate
launch.sh for each console.
- The code adapts to recentlist.json or recentlist-hidden.json, whether
the user chooses to hide the recent panel or not.
- The code rearranges the recentlist itself when launching a game from
the list, a task previously handled by Retroarch.
- Since the new list can contain apps, not just games, a new integer is
assigned to the game structure to track the corresponding line number in
the list.

**Migration:**

- A file named "listMigration.h" is generated to restore the Game
Switcher game list to its previous state. This function parses the
Retroarch list to reconstruct the correct commands in the Miyoo recent
list.
- This file operates independently, facilitating easy removal after some
time.
- It was chosen to fully respect the mainUI logic, even by called the
absolute paths like this :
"/mnt/SDCARD/Emu/GB/../../Roms/GB/Metroid 2 - Return of Samus.gb"
- The migration function also renames old screenshots to adhere to the
new hashes.
- After migration, a flag file is created in
/mnt/SDCARD/Saves/CurrentProfile/config/.recentListMigrated, ensuring
the process occurs only once per profile.

**Keymon :**

A DRASTIC profile is added, configuring the menu button actions and the
proper termination method.
(A button combo is chosen to end Drastic, as SIGTERM was corrupting the
savestates.)

**Runtime :**

The audioserver is initiated, if needed, before launching the Game
Switcher in case the user is coming from a standalone incompatible with.
It now allow to boot on standalones.

**Possibilities :**

Before, games launched by the Game Switcher did not adhere to the
specifics outlined in launch.sh. It is possible to adjust the CPU
frequency for each Retroarch console, with a suggestion to lower it for
Gameboy games to increase gameplay time significantly.

![GameSwitcher_003](https://github.com/OnionUI/Onion/assets/16885275/1214f247-9df2-4154-8719-f44e031d99b6)

---------

Co-authored-by: Schmurtz <Schmurtz@gmail.com>
Co-authored-by: Aemiii91 <44569252+Aemiii91@users.noreply.github.com>
Co-authored-by: Sichroteph <Sichroteph@users.noreply.github.com>
Co-authored-by: schmurtzm <schmurtzm@users.noreply.github.com>