Light filter improvements
- toggle is now a toggle menu icon
- menu item for toggle and intensity are disabled while script is
working to keep the sync (lock file in tmp)
- intensity now changes in realtime & smooth fade
- add hotkey of Menu + B for 2 seconds to enable BLF
- some UI/fading improvements
- tidy up
- persistent across boots if manually enabled in tweaks via "Toggle now"
- Ignore the schedule for the current session if a manual toggle is done
with the hotkey (saves going into tweaks when it's bouncing) The logic
behind this is:
- If the user disables the BLF for some reason while schedule is enabled
we'll skip the schedule for this session, the user can remove this flag
(/tmp/.blfIgnoreSchedule) by flipping the schedule on/off.

- I chose this method of wiping the schedule flag so if the user takes
no action, when they restart the schedule is still on for the next
session.