Add option to disable time restore
This is for the 5 people who are crazy and/or skilled enough to do the
RTC mod.

Normally, Onion saves the system time at shutdown and restores it on
bootup. This disables the restore part.

Since it would be a useless/confusing option for 99.9% of users it has
to be enabled by manually creating the file
`SD:/.tmp_update/config/.noTimeRestore`