Replace global resolution blacklist with a per-game/app whitelis…
Replace the central res_exceptions file, which had a list of strings run
against cmd_to_run.sh, with a per-program approach.
If a game/app wants to launch in 560p, it needs a file `full_resolution`
next to it's launch.sh.

This has multiple benefits, for example

- No need to iterate over a growing file on each game/app launch
- a user can easily remove or add the file if needed

A drawback would be that this is more of a whitelist approach, so for
every system/app/game which can run in 560p, the according file must be
created.