Hide low battery status during screenshot
This commit fixes a small bug where, if the user took a screenshot with
low battery, the low battery icon would be showing in the screenshot.

To fix this, we do the following:

- We have the low battery thread check the raw display value and only
draw if the raw value is not zero
- During the screenshot action, we reset the display to flush the
battery icon out
- We wait a little bit to ensure our display is fresh before
screenshotting

Since the display rawValue gets reset once the screenshot ends, the icon
returns after the screenshot is done.