Create and enable swap partition on boot.
Normally, this is done by MainUI, but when we auto-resume a game, MainUI doesn't get a chance to do it.

MainUI checks if the swap exists already, so no double swapon happens.

Negative impact:

If the cachefile exists (which is the norm) only swapon runs, which is near instant.
If the cachefile doesn't exist, it takes a while to generate the file on SD.